requirements/pytorch-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
argcomplete=1.12.3=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py37hce1f21e_0
brotlipy=0.7.0=py37h5e8e339_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.3.29=h06a4308_1
carla=0.9.12=pypi_0
certifi=2021.10.8=py37h06a4308_2
cffi=1.15.0=py37hd667e15_1
charset-normalizer=2.0.12=pyhd8ed1ab_0
cloudpickle=2.0.0=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cryptography=35.0.0=py37h5d9358c_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=4.4.2=py_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py37_0
expat=2.4.4=h295c915_0
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h6c09931_0
freetype=2.11.0=h70c0345_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
grpcio=1.42.0=py37hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.12.5=pypi_0
gym-carla=0.0.1=dev_0
icu=58.2=he6710b0_3
idna=3.3=pyhd8ed1ab_0
ignite=0.4.8=py_0
imageio=2.10.3=pypi_0
imageio-ffmpeg=0.4.5=pyhd8ed1ab_0
importlib-metadata=4.8.2=py37h06a4308_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.4.1=py37h06a4308_1
ipython=7.31.1=py37h06a4308_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.0.2=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter=1.0.0=py37_7
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.9.1=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.19.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
markupsafe=2.0.1=py37h27cfd23_0
matplotlib=3.4.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
moviepy=1.0.1=py_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.3.0=py37h06a4308_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pypi_0
notebook=6.4.8=py37h06a4308_0
numexpr=2.8.1=py37h6abb31d_0
numpy=1.21.2=py37h20f2e39_0
numpy-base=1.21.2=py37h79a1101_0
olefile=0.46=py37_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1n=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.4=py37h8c16a72_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.4.0=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
proglog=0.1.9=py_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.1=py37h295c915_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.21=pyhd3eb1b0_0
pygame=1.9.6=pypi_0
pyglet=1.5.21=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37h89c1867_4
python=3.7.11=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.10.0=py3.7_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.2.0=pypi_0
pyzmq=22.3.0=py37h295c915_2
qt=5.9.7=h5867ecd_1
qtconsole=5.2.2=pyhd3eb1b0_0
qtpy=1.11.2=pyhd3eb1b0_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd8ed1ab_0
scikit-image=0.16.2=pypi_0
scipy=1.7.2=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=58.0.4=py37h06a4308_0
sip=4.19.8=py37hf484d3e_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.37.2=hc218d9a_0
tensorboard=1.15.0=pyhb230dea_0
terminado=0.13.1=py37h06a4308_0
testpath=0.5.0=pyhd3eb1b0_0
tk=8.6.11=h1ccaba5_0
torchvision=0.11.1=py37_cu102
tornado=6.1=py37h27cfd23_0
tqdm=4.63.0=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.9=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
